+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.55.0] - 2023-07-14
+
+### Changed
+
+- Modified the revenueProp logic to read from eventProperties for Data360 enabled accounts
+- Updated the code to check for existence of eventProperties and access the revenueProp property accordingly
+
 ## [1.50.1] - 2023-06-26
 
 ### Changed